refactor: Separate FormSerialisable classes into separate files
